feat(home): 장소 탐색 지도와 목록 상세 구현
- Place API 기반 지도 마커, 추천 목록, 코스·주차장 상세를 연결\n- 클러스터·현재 위치·주차장 마커와 로그인 게이트를 반영\n- 목록, 상세, 지도 상태 테스트와 Preview 데이터를 보강

+++ feature/home/src/main/java/com/dororong/rodi/feature/home/map/MapClusterer.kt
@@ -20,19 +20,19 @@ data class MapClusterPolicy(
 )
 
 data class ProjectedMapItem(
-    val id: Int,
+    val id: Long,
     val point: GeoPoint,
     val x: Int,
     val y: Int,
 )
 
 data class MapCoursePoint(
-    val id: Int,
+    val id: Long,
     val point: GeoPoint,
 )
 
 data class MapCluster(
-    val memberIds: List<Int>,
+    val memberIds: List<Long>,
     val representativePoint: GeoPoint,
     val targetZoom: Int,
 ) {
